// packages/openapi-typegen/src/openapi/specification.ts
export interface HeaderObject {
  // note: this extends ParameterObject, minus "name" & "in"
  type?: string // required
  description?: string
  required?: boolean
  schema: ReferenceObject | SchemaObject
}

export interface PathItemObject {
  $ref?: string
  summary?: string
  description?: string
  get?: OperationObject
  put?: OperationObject
  post?: OperationObject
  delete?: OperationObject
  options?: OperationObject
  head?: OperationObject
  patch?: OperationObject
  trace?: OperationObject // V3 ONLY
  parameters?: (ReferenceObject | ParameterObject)[]
}

export interface LinkObject {
  operationRef?: string
  operationId?: string
  parameters?: (ReferenceObject | ParameterObject)[]
  requestBody?: RequestBody
  description?: string
}

export interface OperationObject {
  description?: string
  tags?: string[] // unused
  summary?: string // unused
  operationId?: string
  parameters?: (ReferenceObject | ParameterObject)[]
  requestBody?: ReferenceObject | RequestBody
  responses?: Record<string, ReferenceObject | ResponseObject> // required
}

export interface ParameterObject {
  name?: string // required
  in?:
    | 'query'
    | 'header'
    | 'path'
    | /* V3 */ 'cookie'
    | /* V2 */ 'formData'
    | /* V2 */ 'body' // required
  description?: string
  required?: boolean
  deprecated?: boolean
  schema?: ReferenceObject | SchemaObject // required
  type?: 'string' | 'number' | 'integer' | 'boolean' | 'array' | 'file' // V2 ONLY
  items?: ReferenceObject | SchemaObject // V2 ONLY
  enum?: string[] // V2 ONLY
}

export interface ResponseObject {
  description?: string
  headers?: Record<string, ReferenceObject | HeaderObject>
  schema?: ReferenceObject | SchemaObject // V2 ONLY
  links?: Record<string, ReferenceObject | LinkObject> // V3 ONLY
  content?: {
    // V3 ONLY
    [contentType: string]: { schema: ReferenceObject | SchemaObject }
  }
}

export interface RequestBody {
  description?: string
  content?: {
    [contentType: string]: { schema: ReferenceObject | SchemaObject }
  }
}